Crate las

Reads and writes point cloud data stored in the ASPRS las file format.

Reexports

pub use error::Error;

pub use file::File;

pub use point::Point;

pub use stream::Stream;

pub use writer::Writer;

Modules

error
Our errors.

file
las file management.

header
las headers describe the point format, data bounds, and other metadata.

point
las points.

stream
Read a las file like a stream.

vlr
Variable length records.

writer
Write las files.

Type Definitions

Result
Crate-specific resuls.