Struct las::point::Point

pub struct Point {
    pub x: f64,
    pub y: f64,
    pub z: f64,
    pub intensity: u16,
    pub return_number: ReturnNumber,
    pub number_of_returns: NumberOfReturns,
    pub scan_direction: ScanDirection,
    pub edge_of_flight_line: bool,
    pub classification: Classification,
    pub synthetic: bool,
    pub key_point: bool,
    pub withheld: bool,
    pub scan_angle_rank: i8,
    pub user_data: u8,
    pub point_source_id: u16,
    pub gps_time: Option<f64>,
    pub red: Option<u16>,
    pub green: Option<u16>,
    pub blue: Option<u16>,
    pub extra_bytes: Option<Vec<u8>>,
}

A las point.

Fields

x

The x value of the point, as a f64.

y

The y value of the point, as a f64.

z

The z value of the point, as a f64.

intensity

The pulse return magnitude.

This value is system-specific.

return_number

The pulse return number for this point's pulse.

number_of_returns

The total number of returns in this point's pulse.

scan_direction

The scan direction of the mirror, forwards or backwards.

edge_of_flight_line

Is this point at the edge of a flight line?

classification

The classification of this point.

ASPRS defines some integer-to-classification mappings.

synthetic

Was this point created by some other means than LiDAR?

key_point

Is this point a key point?

If so, try not to thin it.

withheld

Should this point be included in processing?

A.k.a. "deleted".

scan_angle_rank

The truncated integer value of the scan angle.

Negative values are left.

user_data

Data used at the user's discretion.

point_source_id

The point source id.

gps_time

The GPS time this point was collected.

Optional, does not exist in all point formats.

red

The red image channel, optional.

green

The green image channel, optional.

blue

The blue image channel, optional.

extra_bytes

Any extra bytes that were included in the point record.

These are legal under the standard, but they break many readers.

Methods

impl Point

fn new() -> Point

Creates a new point.

Examples

use las::point::Point;
let point = Point::new();